docs: add CustomActiveShapePieChart example in TypeScript
